1. 优化在App内存不足情况下销毁界面时，引发一些变量为null的问题

+++ multilibrary/src/main/java/com/zhongjh/albumcamerarecorder/camera/ui/camera/presenter/BaseCameraPicturePresenter.java
@@ -202,7 +202,7 @@ public class BaseCameraPicturePresenter
                 FileUtil.deleteFile(photoFile);
             }
             // 删除多个图片
-            if (photoAdapter.getListData() != null) {
+            if (photoAdapter != null && photoAdapter.getListData() != null) {
                 for (BitmapData bitmapData : photoAdapter.getListData()) {
                     FileUtil.deleteFile(bitmapData.getPath());
                 }
